+++ src/app.ts
@@ -185,22 +185,6 @@ export const Decrypt = (data: string, mode: "Rijndael" | "AES256" = "Rijndael"):
 }
 
 /**
- * To convert string from v1 to v2 https://github.com/darkskyapp/string-hash/blob/master/index.js
- * @param id string
- */ 
-export const ConvertIdFromV1ToV2 = (str: any) =>  {
-    let hash = 5381,
-    i = str.length;
-    while(i) {
-      hash = (hash * 33) ^ str.charCodeAt(--i);
-    }
-    /* JavaScript does bitwise operations (like XOR, above) on 32-bit signed
-    * integers. Since we want the results to be always positive, convert the
-    * signed int to an unsigned by doing an unsigned bitshift. */
-    return hash >>> 0;
-} 
-
-/**
  *
  */
 export const Download = function (url: string): Promise<Buffer> {
